rearranged plotting functions in different files for amin functions and helper functions, adapted __init__
