+++ petab/visualize/__init__.py
@@ -1,5 +1,15 @@
-from .get_data_to_plot import *  # noqa: F403, F401
-from .plot_data_and_simulation import *  # noqa: F403, F401
-from .plotting_config import *  # noqa: F403, F401
-from .plot_measurement import *  # noqa: F403, F401
-from .plot_simulation import *  # noqa: F403, F401
+"""
+Visualize
+=========
+
+PEtab comes with visualization functionality. Those need to be imported via
+import petab.visualize.
+
+"""
+
+from .plot_data_and_simulation import (plot_data_and_simulation,
+                                       plot_measurements_by_observable)
+
+__all__ = ["plot_data_and_simulation",
+           "plot_measurements_by_observable"
+           ]
